Use XHR instead of JSONP
Finally set up lighttpd so I can do this the right way.

Also put the raw lists of relations into their own files, with a script to
automatically rebuild all the data. Saves a ton of time when I make data format
changes like now.

Put in custom lat-lon for New York and Chicago (both of which are quite
asymmetric and off-center in their layout).