Ensure consistent random color generation
This is why the New York data inexplicably changed in the last commit -- the
shuttles have no color, and they got set to new random colors every time I
rebuilt. I've rebuilt with the consistent coloring, so it should be fine now.

Also I discovered it's handy to bust the browser's cache of the data files when
developing, so every load is a full reload.