Add "local data" concept, use it to unshittify Tokyo data
I downloaded data from the Ministry of Land, Infrastructure and Transportation,
which has a bunch of railway data. I pulled the data for the lines that are
missing or mostly missing, and now the Tokyo map looks way better.

I decided to store the local data as OSM-formatted XML files, so they can be
read by the existing code.

I'll check in the script I was using to process the MLIT data file once I clean
it up. If I get bored I might do the data for some other lines (maybe Rinkai,
Yamanote).